pyproject.toml
[project]
name = "cnn-quantization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "omegaconf>=2.3.0",
    "pandas>=3.0.2",
    "seaborn>=0.13.2",
    "torch>=2.1.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
]
